Fix Stack ID input bug & add Stack ID history
Fixed issue with input field not moving to the next input automatically and therefore not running the auto retrive function.

Also added small view for Loaders to see the Stacks that have already being retrieved. Also added logic to make sure the same stackID could not be retrieve and potentailly added to the container preventing duplicated data.