Implement delete function and tidy directory
First I had to keep the stackID on the stacks when recalled from storage and put into the container. I also had to assign a stackId to any stacks that were constructed from the productlist and placed straight into the container. This was so that I could select and delete stacks based on their stackID.

I also cleaned up the ContanierPreview in order to make the component more readable.